+++ provider/s3/s3_test.go
@@ -58,8 +58,7 @@ func TestS3_Load(t *testing.T) {
 
 			loader := ks3.New(
 				"bucket/key",
-				ks3.WithAWSConfig(cfg),
-				ks3.WithUnmarshal(testcase.unmarshal),
+				append(testcase.opts, ks3.WithAWSConfig(cfg))...,
 			)
 			values, err := loader.Load()
 			if testcase.err != "" {
@@ -75,7 +74,7 @@ func TestS3_Load(t *testing.T) {
 func TestS3_Watch(t *testing.T) {
 	t.Parallel()
 
-	for _, testcase := range testcases() {
+	for _, testcase := range append(testcases(), watchcases()...) {
 		testcase := testcase
 
 		t.Run(testcase.description, func(t *testing.T) {
@@ -100,9 +99,7 @@ func TestS3_Watch(t *testing.T) {
 			var err atomic.Pointer[error]
 			loader := ks3.New(
 				"bucket/key",
-				ks3.WithAWSConfig(cfg),
-				ks3.WithPollInterval(10*time.Millisecond),
-				ks3.WithUnmarshal(testcase.unmarshal),
+				append(testcase.opts, ks3.WithAWSConfig(cfg))...,
 			)
 			loader.Status(func(_ bool, e error) {
 				if e != nil {
@@ -124,6 +121,14 @@ func TestS3_Watch(t *testing.T) {
 				assert.NoError(t, e)
 			}()
 			<-started
+			if testcase.event != nil {
+				err := loader.OnEvent(testcase.event)
+				if testcase.err != "" {
+					assert.EqualError(t, err, testcase.err)
+				} else {
+					assert.NoError(t, err)
+				}
+			}
 
 			time.Sleep(15 * time.Millisecond) // wait for the first tick, but not the second
 			select {
@@ -138,30 +143,26 @@ func TestS3_Watch(t *testing.T) {
 	}
 }
 
-func testcases() []struct {
+type testcase struct {
 	description string
+	opts        []ks3.Option
+	event       []byte
 	middleware  func(
 		context.Context,
 		middleware.FinalizeInput,
 		middleware.FinalizeHandler,
 	) (middleware.FinalizeOutput, middleware.Metadata, error)
-	unmarshal func([]byte, any) error
-	expected  map[string]any
-	err       string
-} {
-	return []struct {
-		description string
-		middleware  func(
-			context.Context,
-			middleware.FinalizeInput,
-			middleware.FinalizeHandler,
-		) (middleware.FinalizeOutput, middleware.Metadata, error)
-		unmarshal func([]byte, any) error
-		expected  map[string]any
-		err       string
-	}{
+	expected map[string]any
+	err      string
+}
+
+func testcases() []testcase {
+	return []testcase{
 		{
-			description: "s3",
+			description: "pull only",
+			opts: []ks3.Option{
+				ks3.WithPollInterval(10 * time.Millisecond),
+			},
 			middleware: func(
 				ctx context.Context,
 				_ middleware.FinalizeInput,
@@ -185,6 +186,9 @@ func testcases() []struct {
 		},
 		{
 			description: "get object error",
+			opts: []ks3.Option{
+				ks3.WithPollInterval(10 * time.Millisecond),
+			},
 			middleware: func(
 				ctx context.Context,
 				_ middleware.FinalizeInput,
@@ -201,6 +205,12 @@ func testcases() []struct {
 		},
 		{
 			description: "unmarshal error",
+			opts: []ks3.Option{
+				ks3.WithPollInterval(10 * time.Millisecond),
+				ks3.WithUnmarshal(func([]byte, any) error {
+					return errors.New("unmarshal error")
+				}),
+			},
 			middleware: func(
 				ctx context.Context,
 				_ middleware.FinalizeInput,
@@ -218,10 +228,131 @@ func testcases() []struct {
 					return middleware.FinalizeOutput{}, middleware.Metadata{}, nil
 				}
 			},
-			unmarshal: func([]byte, any) error {
-				return errors.New("unmarshal error")
-			},
 			err: "unmarshal: unmarshal error",
+		},
+	}
+}
+
+func watchcases() []testcase {
+	return []testcase{
+		{
+			description: "object created (sns)",
+			event: []byte(`
+{
+   "Records":[
+      {
+         "eventSource":"aws:s3",
+         "eventName":"ObjectCreated:Put",
+         "s3":{
+            "bucket":{
+               "name":"bucket"
+            },
+            "object":{
+               "key":"key"
+            }
+         }
+      }
+   ]
+}`),
+			middleware: func(
+				ctx context.Context,
+				_ middleware.FinalizeInput,
+				_ middleware.FinalizeHandler,
+			) (middleware.FinalizeOutput, middleware.Metadata, error) {
+				switch awsMiddleware.GetOperationName(ctx) {
+				case "GetObject":
+					return middleware.FinalizeOutput{
+						Result: &s3.GetObjectOutput{
+							Body: io.NopCloser(strings.NewReader(`{"k":"v"}`)),
+							ETag: aws.String("k42"),
+						},
+					}, middleware.Metadata{}, nil
+				default:
+					return middleware.FinalizeOutput{}, middleware.Metadata{}, nil
+				}
+			},
+			expected: map[string]any{
+				"k": "v",
+			},
+		},
+		{
+			description: "object created (event bridge)",
+			event: []byte(`
+{
+  "detail-type": "Object Created",
+  "source": "aws.s3",
+  "detail": {
+    "bucket": {
+      "name": "bucket"
+    },
+    "object": {
+      "key": "key"
+    }
+  }
+}`),
+			middleware: func(
+				ctx context.Context,
+				_ middleware.FinalizeInput,
+				_ middleware.FinalizeHandler,
+			) (middleware.FinalizeOutput, middleware.Metadata, error) {
+				switch awsMiddleware.GetOperationName(ctx) {
+				case "GetObject":
+					return middleware.FinalizeOutput{
+						Result: &s3.GetObjectOutput{
+							Body: io.NopCloser(strings.NewReader(`{"k":"v"}`)),
+							ETag: aws.String("k42"),
+						},
+					}, middleware.Metadata{}, nil
+				default:
+					return middleware.FinalizeOutput{}, middleware.Metadata{}, nil
+				}
+			},
+			expected: map[string]any{
+				"k": "v",
+			},
+		},
+		{
+			description: "unmatched message created",
+			event: []byte(`
+{
+   "Records":[
+      {
+         "eventSource":"aws:s3",
+         "eventName":"s3:ObjectCreated:Put",
+         "s3":{
+            "bucket":{
+               "name":"another-bucket"
+            },
+            "object":{
+               "key":"key"
+            }
+         }
+      }
+   ]
+}`),
+			middleware: func(
+				context.Context,
+				middleware.FinalizeInput,
+				middleware.FinalizeHandler,
+			) (middleware.FinalizeOutput, middleware.Metadata, error) {
+				return middleware.FinalizeOutput{}, middleware.Metadata{}, nil
+			},
+			expected: map[string]any{
+				"k": "v",
+			},
+			err: "unsupported s3 event: unsupported operation",
+		},
+		{
+			description: "no-json messages",
+			event:       []byte(`not a json`),
+			middleware: func(
+				context.Context,
+				middleware.FinalizeInput,
+				middleware.FinalizeHandler,
+			) (middleware.FinalizeOutput, middleware.Metadata, error) {
+				return middleware.FinalizeOutput{}, middleware.Metadata{}, nil
+			},
+			err: "unmarshal s3 event: invalid character 'o' in literal null (expecting 'u')",
 		},
 	}
 }