+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Support change notification via SNS topic (#267).
+
 ## [1.0.0] - 2024-03-16
 
 First stable release.